+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Log the system order on the dispatcher (#414, **@RiscadoA**).
+- Negative tags to the dispatcher (#390, **@RiscadoA**).
 - Support for queries with multiple relation terms (#929, **@RiscadoA**).
 - Support for queries with more than one unrelated targets (#930, **@RiscadoA**).
 - Fixed Time Step plugin (#989, **@joaomanita**).